Merged PR 29418: Add modelAsExtensible to the SDK's Enums
adding modelAsExtensible will enable our SDK users to get the "enum" experience (instead of using hard coded strings) in most of the cases. since the AutoReset generates readonly strings it can't be used as a switch case values nor as optional parameters in function definition.
This is a breaking change and **should not be** "cherry-pick"ed to V2 branch