Merged PR 220055: Fix userId parameter in subscriptions API
Fix userId parameter in subscriptions API

Fix bug with the C# SDK with GetUserSubscriptionsAsAdmin. Whentrying to call this method with the GUID for a userGraphId, it fails because it does not correctly parse the userGraphId into the URL. Here you can see the difference between calling GetUserArtifactAccessAsAdmin which works correctly and GetUserSubscriptionsAsAdmin which fails.
 ![image.png](https://powerbi.visualstudio.com/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/4f4b7330-5c73-4eb2-b9cb-d2f4c16debf7/pullRequests/220055/attachments/image.png)