Merged PR 254682: Add limitation in RefreshDataset and RefreshDatasetInGroup
We have an incident where the customer is unable to do a full refresh of an IR table. They’re using the Power BI Refresh REST API and are providing the applyRefreshPolicy = false on their request and are complaining that their request is not having the desired effect. **(Actually, the root cause is that the customer’s dataset is on Shared while we only support this operation in Premium)**
We add limitation statement in refreshDataset and refreshDatasetInGroup to make it clear.
Preview: https://review.docs.microsoft.com/en-us/rest/api/power-bi/datasets/refresh-dataset?branch=yuhsienchen%2Fadd-limitation-description
https://review.docs.microsoft.com/en-us/rest/api/power-bi/datasets/refresh-dataset-in-group?branch=yuhsienchen%2Fadd-limitation-description